update: #686: `Parser::parse`と`Parser::parse_blocking`の実装を`src/parser/parse.rs`に移動
